Получение данных из СУБД (#12)
* Убран лишний textEdit в интерфейсе менеджера

* Убран лишний код в интерфейсе менеджера

* Вынесены некоторые методы за try в интерфейсе менеджера

* Улучшена надпись с приветствием

* Почищен код интерфейса руководителя

* Перенесен код в интерфейс, почищен интерфейс и код

* Добавлена работа с уже готовыми функциями СУБД для руководителя

* Исправлены названия элементов интерфейса

* Улучшен интерфейс продавца

* Добавлена работа с СУБД, убрано лишнее в интерфейсе продавца

* Перенесен код в интерфейс, убрано лишнее в интерфейсе маркетолога

* Добавлено получение сотрудников из СУБД

* Все названия столбцов сделаны константами

* Добавлено получение данных о рекламе из СУБД

* Добавлена работа с СУБД в интерфейс маркетолога

* Исправления интерфейса окон

* Исправления в типах СУБД

* Добавлено получение списка клиентов из СУБД

* Добавлена вкладка с клиентами для руководителя

* Мелкие исправления в интерфейсе

* Мелкие улучшения

* Исправлен тип для соответствия типу СУБД

+++ CRM/database-types/advertisement.h
@@ -22,7 +22,7 @@
 struct advertisement {
     int id;                // ID рекламы
     std::string source;    // Источник рекламы
-    long long budget;      // Бюджет рекламы
+    int budget;            // Бюджет рекламы
     std::string comments;  // Комментарий к рекламе
 
     advertisement() = default;
@@ -61,7 +61,7 @@ struct type_conversion<advertisement> {
         try {
             ad.id = v.get<int>("id", 0);
             ad.source = v.get<std::string>("source", "");
-            ad.budget = v.get<long long>("budget", 0);
+            ad.budget = v.get<int>("budget", 0);
             ad.comments = v.get<std::string>("comments", "");
         } catch (std::exception const& e) {
             // Logging
